Fixed more size_t
git-svn-id: http://openmesh.org/svnrepo/OpenMesh@858 fdac6126-5c0c-442c-9429-916003d36597

+++ trunk/src/OpenMesh/Tools/Utils/HeapT.hh
@@ -311,10 +311,9 @@ HeapT<HeapEntry, HeapInterface>::
 upheap(unsigned int _idx)
 {
   HeapEntry     h = entry(_idx);
-  unsigned int  parentIdx;
+  size_t        parentIdx;
 
-  while ((_idx>0) &&
-	 interface_.less(h, entry(parentIdx=parent(_idx))))
+  while ((_idx>0) && interface_.less(h, entry(parentIdx=parent(_idx))))
   {
     entry(_idx, entry(parentIdx));
     _idx = parentIdx;
